requirements.txt
pytz==2024.2
speedtest-cli==2.1.3
